voice-chat-bot/requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
ffmpeg
PyNaCl
youtube-dl
yt_dlp
